score/locomotives/requirements.txt
pip==22.2.2
setuptools==63.2.0
wheel==0.37.1
django==4.1.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1
gunicorn==20.1.0
django-cors-headers==3.13.0
scipy
cvxopt
celery